# frontend/openapi.yaml
openapi: 3.1.0
info:
  title: Timverse
  version: 0.1.0
paths:
...
  /api/v1/specimens/:
    get:
      tags:
        - specimens
      summary: Read Specimens
      description: Retrieve specimens.
      operationId: specimens-read_specimens
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            title: Limit
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimensPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      tags:
        - specimens
      summary: Create Specimen
      description: Submit a new specimen for review.
      operationId: specimens-create_specimen
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingSpecimenCreate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/v1/specimens/{id}:
    get:
      tags:
        - specimens
      summary: Read Specimen
      description: Get specimen by ID.
      operationId: specimens-read_specimen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    put:
      tags:
        - specimens
      summary: Update Specimen
      description: Submit an update for an existing specimen.
      operationId: specimens-update_specimen
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingSpecimenUpdate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - specimens
      summary: Delete Specimen
      description: Delete a specimen.
      operationId: specimens-delete_specimen
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Specimens-Delete Specimen
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/v1/pending-specimens/{pending_id}/approve:
    post:
      tags:
        - pending-specimens
      summary: Approve Pending Specimen
      description: Approve a pending specimen and apply it to the specimen table.
      operationId: pending-specimens-approve_pending_specimen
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: pending_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Pending Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingSpecimenReview"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/pending-specimens/{pending_id}/reject:
    post:
      tags:
        - pending-specimens
      summary: Reject Pending Specimen Route
      description: Reject a pending specimen without touching the specimen table.
      operationId: pending-specimens-reject_pending_specimen_route
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: pending_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Pending Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingSpecimenReview"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/pending-specimens/{pending_id}:
    put:
      tags:
        - pending-specimens
      summary: Update Pending Specimen
      description: |-
        Update the changed_data of an existing pending specimen
        before it is approved/rejected.
      operationId: pending-specimens-update_pending_specimen
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: pending_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Pending Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingSpecimenUpdate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - pending-specimens
      summary: Delete Pending Specimen
      description: Delete a pending specimen (only allowed while status is PENDING).
      operationId: pending-specimens-delete_pending_specimen
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: pending_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Pending Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimenPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/pending-specimens/:
    get:
      tags:
        - pending-specimens
      summary: List Pending Specimens
      description: |-
        List pending specimens, optionally filtered by status.
        If no status is provided, all pending specimens are returned.
      operationId: pending-specimens-list_pending_specimens
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: "#/components/schemas/PendingStatus"
              - type: "null"
            title: Status
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimensPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/pending-specimens/specimen/{specimen_id}:
    get:
      tags:
        - pending-specimens
      summary: List Approved Specimen Trail
      description: List all pending specimens.
      operationId: pending-specimens-list_approved_specimen_trail
      parameters:
        - name: specimen_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Specimen Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSpecimensPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
    AssemblyType:
      type: string
      enum:
        - Wall-Floor
        - Wall-Wall
        - Wall-Floor & Wall-Wall
      title: AssemblyType
...
    DOI:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        link:
          type: string
          maxLength: 255
          minLength: 1
          title: Link
        ref_title:
          type: string
          maxLength: 255
          minLength: 1
          title: Ref Title
        authors:
          type: string
          maxLength: 255
          minLength: 1
          title: Authors
        pub_year:
          type: integer
          title: Pub Year
      type: object
      required:
        - link
        - ref_title
        - authors
        - pub_year
      title: DOI
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JoineryType:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
        has_dowel:
          type: boolean
          title: Has Dowel
      type: object
      required:
        - label
        - has_dowel
      title: JoineryType
...
    PendingSpecimenCreate:
      properties:
        comment_by_author:
          anyOf:
            - type: string
            - type: "null"
          title: Comment By Author
        specimen_reference_id:
          type: string
          title: Specimen Reference Id
        replicate_tests:
          type: integer
          title: Replicate Tests
        note:
          anyOf:
            - type: string
            - type: "null"
          title: Note
        dowel:
          type: boolean
          title: Dowel
        connector:
          type: boolean
          title: Connector
        assembly_type:
          $ref: "#/components/schemas/AssemblyType"
        practice:
          $ref: "#/components/schemas/Practice"
        connection_description:
          anyOf:
            - type: string
            - type: "null"
          title: Connection Description
        element_dimension:
          type: string
          maxLength: 1000
          title: Element Dimension
        fastener_numbers:
          type: integer
          title: Fastener Numbers
        moisture_percentage:
          type: string
          title: Moisture Percentage
        wood_type:
          anyOf:
            - type: string
            - type: "null"
          title: Wood Type
        wood_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Wood Mechanical Properties
        connector_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Connector Mechanical Properties
        fastener_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Fastener Mechanical Properties
        e_date:
          anyOf:
            - type: string
            - type: "null"
          title: E Date
        e_test_loading_type:
          anyOf:
            - $ref: "#/components/schemas/TestLoadingType"
            - type: "null"
        e_yield_point_method:
          anyOf:
            - $ref: "#/components/schemas/YieldPointMethod"
            - type: "null"
        e_stiffness:
          anyOf:
            - type: number
            - type: "null"
          title: E Stiffness
        e_yield_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Yield Displacement
        e_yield_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Yield Force
        e_max_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Max Displacement
        e_max_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Max Force
        e_ultimate_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Ultimate Displacement
        e_ultimate_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Ultimate Force
        e_ductility:
          anyOf:
            - type: number
            - type: "null"
          title: E Ductility
        e_qfm_description:
          anyOf:
            - type: string
            - type: "null"
          title: E Qfm Description
        doi_id:
          type: string
          format: uuid
          title: Doi Id
        e_qualitative_failure_measure:
          items:
            type: string
            format: uuid
          type: array
          title: E Qualitative Failure Measure
        fastener_type_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Fastener Type Ids
        loading_direction_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Loading Direction Ids
        joinery_type_id:
          type: string
          format: uuid
          title: Joinery Type Id
        sub_joinery_type_id:
          type: string
          format: uuid
          title: Sub Joinery Type Id
      type: object
      required:
        - specimen_reference_id
        - replicate_tests
        - dowel
        - connector
        - assembly_type
        - practice
        - element_dimension
        - fastener_numbers
        - moisture_percentage
        - doi_id
        - e_qualitative_failure_measure
        - fastener_type_ids
        - loading_direction_ids
        - joinery_type_id
        - sub_joinery_type_id
      title: PendingSpecimenCreate
      description: Body for submitting a brand new specimen for review.
    PendingSpecimenPublic:
      properties:
        comment_by_author:
          anyOf:
            - type: string
            - type: "null"
          title: Comment By Author
        specimen_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Specimen Id
        changed_by_user_id:
          type: string
          format: uuid
          title: Changed By User Id
        changed_data:
          additionalProperties: true
          type: object
          title: Changed Data
        id:
          type: string
          format: uuid
          title: Id
        status:
          $ref: "#/components/schemas/PendingStatus"
        created_at:
          type: string
          format: date-time
          title: Created At
        reviewer_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Reviewer Id
        comment_by_reviewer:
          anyOf:
            - type: string
            - type: "null"
          title: Comment By Reviewer
        reviewed_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Reviewed At
      type: object
      required:
        - changed_by_user_id
        - changed_data
        - id
        - status
        - created_at
      title: PendingSpecimenPublic
    PendingSpecimenReview:
      properties:
        comment_by_reviewer:
          type: string
          title: Comment By Reviewer
      type: object
      required:
        - comment_by_reviewer
      title: PendingSpecimenReview
      description: Body for approve or reject actions.
    PendingSpecimenUpdate:
      properties:
        comment_by_author:
          anyOf:
            - type: string
            - type: "null"
          title: Comment By Author
        specimen_reference_id:
          anyOf:
            - type: string
            - type: "null"
          title: Specimen Reference Id
        replicate_tests:
          anyOf:
            - type: integer
            - type: "null"
          title: Replicate Tests
        note:
          anyOf:
            - type: string
            - type: "null"
          title: Note
        dowel:
          anyOf:
            - type: boolean
            - type: "null"
          title: Dowel
        connector:
          anyOf:
            - type: boolean
            - type: "null"
          title: Connector
        assembly_type:
          anyOf:
            - $ref: "#/components/schemas/AssemblyType"
            - type: "null"
        practice:
          anyOf:
            - $ref: "#/components/schemas/Practice"
            - type: "null"
        connection_description:
          anyOf:
            - type: string
            - type: "null"
          title: Connection Description
        element_dimension:
          anyOf:
            - type: string
            - type: "null"
          title: Element Dimension
        fastener_numbers:
          anyOf:
            - type: integer
            - type: "null"
          title: Fastener Numbers
        moisture_percentage:
          anyOf:
            - type: string
            - type: "null"
          title: Moisture Percentage
        wood_type:
          anyOf:
            - type: string
            - type: "null"
          title: Wood Type
        wood_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Wood Mechanical Properties
        connector_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Connector Mechanical Properties
        fastener_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Fastener Mechanical Properties
        e_date:
          anyOf:
            - type: string
            - type: "null"
          title: E Date
        e_test_loading_type:
          anyOf:
            - $ref: "#/components/schemas/TestLoadingType"
            - type: "null"
        e_yield_point_method:
          anyOf:
            - $ref: "#/components/schemas/YieldPointMethod"
            - type: "null"
        e_stiffness:
          anyOf:
            - type: number
            - type: "null"
          title: E Stiffness
        e_yield_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Yield Displacement
        e_yield_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Yield Force
        e_max_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Max Displacement
        e_max_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Max Force
        e_ultimate_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Ultimate Displacement
        e_ultimate_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Ultimate Force
        e_ductility:
          anyOf:
            - type: number
            - type: "null"
          title: E Ductility
        e_qfm_description:
          anyOf:
            - type: string
            - type: "null"
          title: E Qfm Description
        e_qualitative_failure_measure:
          anyOf:
            - items:
                type: string
                format: uuid
              type: array
            - type: "null"
          title: E Qualitative Failure Measure
        fastener_type_ids:
          anyOf:
            - items:
                type: string
                format: uuid
              type: array
            - type: "null"
          title: Fastener Type Ids
        loading_direction_ids:
          anyOf:
            - items:
                type: string
                format: uuid
              type: array
            - type: "null"
          title: Loading Direction Ids
        joinery_type_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Joinery Type Id
        sub_joinery_type_id:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Sub Joinery Type Id
      type: object
      title: PendingSpecimenUpdate
      description: Body for submitting or editing specimen changes under review.
    PendingSpecimensPublic:
      properties:
        pending_specimens:
          items:
            $ref: "#/components/schemas/PendingSpecimenPublic"
          type: array
          title: Pending Specimens
        count:
          type: integer
          title: Count
      type: object
      required:
        - pending_specimens
        - count
      title: PendingSpecimensPublic
      description: Response model for listing multiple pending specimens.
    PendingStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
      title: PendingStatus
    Practice:
      type: string
      enum:
        - Conventional
        - Research and Development
      title: Practice
...
    SpecimenPublic:
      properties:
        specimen_reference_id:
          type: string
          title: Specimen Reference Id
        replicate_tests:
          type: integer
          title: Replicate Tests
        note:
          anyOf:
            - type: string
            - type: "null"
          title: Note
        dowel:
          type: boolean
          title: Dowel
        connector:
          type: boolean
          title: Connector
        assembly_type:
          $ref: "#/components/schemas/AssemblyType"
        practice:
          $ref: "#/components/schemas/Practice"
        connection_description:
          anyOf:
            - type: string
            - type: "null"
          title: Connection Description
        element_dimension:
          type: string
          maxLength: 1000
          title: Element Dimension
        fastener_numbers:
          type: integer
          title: Fastener Numbers
        moisture_percentage:
          type: string
          title: Moisture Percentage
        wood_type:
          anyOf:
            - type: string
            - type: "null"
          title: Wood Type
        wood_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Wood Mechanical Properties
        connector_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Connector Mechanical Properties
        fastener_mechanical_properties:
          anyOf:
            - type: string
            - type: "null"
          title: Fastener Mechanical Properties
        e_date:
          anyOf:
            - type: string
            - type: "null"
          title: E Date
        e_test_loading_type:
          anyOf:
            - $ref: "#/components/schemas/TestLoadingType"
            - type: "null"
        e_yield_point_method:
          anyOf:
            - $ref: "#/components/schemas/YieldPointMethod"
            - type: "null"
        e_stiffness:
          anyOf:
            - type: number
            - type: "null"
          title: E Stiffness
        e_yield_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Yield Displacement
        e_yield_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Yield Force
        e_max_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Max Displacement
        e_max_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Max Force
        e_ultimate_displacement:
          anyOf:
            - type: number
            - type: "null"
          title: E Ultimate Displacement
        e_ultimate_force:
          anyOf:
            - type: number
            - type: "null"
          title: E Ultimate Force
        e_ductility:
          anyOf:
            - type: number
            - type: "null"
          title: E Ductility
        e_qfm_description:
          anyOf:
            - type: string
            - type: "null"
          title: E Qfm Description
        id:
          type: string
          format: uuid
          title: Id
        uploader_id:
          type: string
          format: uuid
          title: Uploader Id
        doi:
          $ref: "#/components/schemas/DOI"
        joinery_type:
          $ref: "#/components/schemas/JoineryType"
        sub_joinery_type:
          $ref: "#/components/schemas/SubJoineryType"
        e_qualitative_failure_measure:
          items:
            $ref: "#/components/schemas/FailureMode"
          type: array
          title: E Qualitative Failure Measure
        fastener_types:
          items:
            $ref: "#/components/schemas/FastenerType"
          type: array
          title: Fastener Types
        loading_directions:
          items:
            $ref: "#/components/schemas/LoadingDirection"
          type: array
          title: Loading Directions
      type: object
      required:
        - specimen_reference_id
        - replicate_tests
        - dowel
        - connector
        - assembly_type
        - practice
        - element_dimension
        - fastener_numbers
        - moisture_percentage
        - id
        - uploader_id
        - doi
        - joinery_type
        - sub_joinery_type
        - e_qualitative_failure_measure
        - fastener_types
        - loading_directions
      title: SpecimenPublic
    SpecimensPublic:
      properties:
        data:
          items:
            $ref: "#/components/schemas/SpecimenPublic"
          type: array
          title: Data
        count:
          type: integer
          title: Count
      type: object
      required:
        - data
        - count
      title: SpecimensPublic
    SubJoineryType:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        joinery_type_id:
          type: string
          format: uuid
          title: Joinery Type Id
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
      type: object
      required:
        - joinery_type_id
        - label
      title: SubJoineryType
...
    TestLoadingType:
      type: string
      enum:
        - Cyclic
        - Monotonic
        - Monotonic and Cyclic
      title: TestLoadingType
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    YieldPointMethod:
      type: string
      enum:
        - CEN 1/6
        - EEEP
        - FEMA P795
      title: YieldPointMethod
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/login/access-token